refactor(#9): add generate token method to utils
