+++ lib/utils.js
@@ -1,4 +1,7 @@
 import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
+
+import { jwtSecret, jwtExpiresIn } from '../config/keys.js';
 
 export const hashPassword = async password => {
   if (typeof password !== 'string' || password.length === 0) {
@@ -11,4 +14,20 @@ export const hashPassword = async password => {
 
 export const comparePassword = async (plainPassword, hashedPassword) => {
   return await bcrypt.compare(plainPassword, hashedPassword);
+};
+
+export const generateToken = user => {
+  const token = jwt.sign(
+    {
+      id: user.id,
+      fullName: user.fullName,
+      email: user.email,
+      role: user.role
+    },
+    jwtSecret,
+    {
+      expiresIn: jwtExpiresIn
+    }
+  );
+  return token;
 };
